change cache retrieving method and move cache test to different file
